# smartgas-payments-composer/backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server celler server.",
        "title": "Smart Gas API Specification",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/synchronizations": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get paginated synchronizations",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Synchronization"
                ],
                "summary": "Synchronization List",
                "parameters": [
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "gas_pumps",
                            "gas_stations",
                            "customer_levels"
                        ],
                        "type": "string",
                        "name": "type",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Synchronizations List",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.PaginationResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/dto.SynchronizationListResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request, failed on body, form, query...",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.BadRequestMessage"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.GeneralMessage"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.GeneralMessage"
                        }
                    }
                }
            }
        },
        "/api/v1/synchronizations/last": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get last synchronization",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Synchronization"
                ],
                "summary": "Last Synchronization",
                "parameters": [
                    {
                        "enum": [
                            "gas_pumps",
                            "gas_stations",
                            "customer_levels"
                        ],
                        "type": "string",
                        "name": "type",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Access token \u0026 Refresh Token",
                        "schema": {
                            "$ref": "#/definitions/dto.SynchronizationGetLastSyncResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request, failed on body, form, query...",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.BadRequestMessage"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.GeneralMessage"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.GeneralMessage"
                        }
                    }
                }
            }
        },
        "/api/v1/synchronizations/now": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Synchronize gas pumps, gas stations or all",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Synchronization"
                ],
                "summary": "Sync Now",
                "parameters": [
                    {
                        "description": "Sync type",
                        "name": "params",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.SynchronizationNowQueryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Synchronized",
                        "schema": {
                            "$ref": "#/definitions/dto.GeneralMessage"
                        }
                    },
                    "400": {
                        "description": "Bad Request, failed on body, form, query...",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.BadRequestMessage"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.GeneralMessage"
                        }
                    },
                    "423": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.GeneralMessage"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.GeneralMessage"
                        }
                    }
                }
            }
        },
        "/api/v1/synchronizations/{id}/details": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get paginated synchronizations detail List",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Synchronization"
                ],
                "summary": "Synchronization Detail List",
                "parameters": [
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "maxLength": 36,
                        "minLength": 36,
                        "type": "string",
                        "description": "uuid4 id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Synchronizations Detail List",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.PaginationResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/dto.SynchronizationListDetailResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request, failed on body, form, query...",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.BadRequestMessage"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.GeneralMessage"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/dto.GeneralMessage"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.BadRequestMessage": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "dto.GeneralMessage": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string"
                }
            }
        },
...
        "dto.PaginationResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "format": "array"
                },
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                },
                "total_rows": {
                    "type": "integer"
                }
            }
        },
...
        "dto.SynchronizationGetLastSyncResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "dto.SynchronizationListDetailResponse": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "data": {
                    "type": "string"
                },
                "error_text": {
                    "type": "string"
                },
                "external_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                }
            }
        },
        "dto.SynchronizationListResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "text": {
                                "type": "string"
                            }
                        }
                    }
                },
                "id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "dto.SynchronizationNowQueryRequest": {
            "type": "object",
            "required": [
                "type"
            ],
            "properties": {
                "type": {
                    "type": "string",
                    "enum": [
                        "gas_pumps",
                        "gas_stations",
                        "customer_levels"
                    ]
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" followed by a space and paste the access token",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}